uses bcrypt
