+++ dashboard/15-final/middleware.ts
@@ -1,5 +1,7 @@
-export { auth as middleware } from './auth';
+import NextAuth from 'next-auth';
+import { authConfig } from './auth.config';
+export default NextAuth(authConfig).auth;
 
 export const config = {
-  matcher: ['/((?!_next/static|_next/image|favicon.ico).*)'],
+  matcher: ['/((?!api|_next/static|_next/image|favicon.ico).*)'],
 };
