+++ dashboard/15-final/package.json
@@ -15,7 +15,7 @@
     "@types/react-dom": "18.2.7",
     "@vercel/postgres": "^0.5.0",
     "autoprefixer": "10.4.16",
-    "bcryptjs": "^2.4.3",
+    "bcrypt": "^5.1.1",
     "clsx": "^2.0.0",
     "next": "13.5.7-canary.23",
     "next-auth": "5.0.0-beta.0",
@@ -28,7 +28,7 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.5",
+    "@types/bcrypt": "^5.0.1",
     "dotenv": "^16.3.1",
     "prettier": "^3.0.3"
   },
